Revert "Merge pull request #29 from NachiketKandari/db-migration"
This reverts commit 4d935a08fe26a37835e19cc949f8d67259ddb68d, reversing
changes made to e1f43a2130d1aa3b32cc130d84b2c7f5c1f48600.

+++ backend/src/insightxpert/training/schema.py
@@ -6,7 +6,7 @@ CREATE TABLE transactions (
     timestamp TEXT,
     transaction_type TEXT,
     merchant_category TEXT,
-    amount_inr DOUBLE PRECISION,
+    amount_inr REAL,
     transaction_status TEXT,
     sender_age_group TEXT,
     receiver_age_group TEXT,
